+++ phon/utils/text.hpp
@@ -30,7 +30,12 @@ Array<Array<String>> parse_csv(const String &path, std::string_view splitter = "
 
 void write_csv(const String &path, const Array<Array<String>> &csv, std::string_view separator = ",");
 
+Array<double> read_matrix(const String &path, std::string_view splitter = ",");
+
+void write_matrix(const Array<double> &matrix, const String &path, std::string_view sep = ",");
+
 String matrix_to_csv(const Array<double> &matrix, std::string_view sep = ",");
+
 
 }} // namespace phonometrica::utils