+++ docs/about/release-notes.rst
@@ -1,6 +1,12 @@
 Release notes
 -------------
 
+0.7.2 (??/11/2019)
+~~~~~~~~~~~~~~~~~~
+
+- ``lm()`` function for linear regression
+- ``read_matrix()`` and ``write_matrix()`` functions to read/write a numeric array to/from a text file
+
 0.7.1 (07/11/2019)
 ~~~~~~~~~~~~~~~~~~